third_party/blink/web_tests/fast/events/wheel/wheelevent-handler-count-expected.txt

This test checks that we correctly update the wheel event handler count as event handlers are added and removed

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS internals.wheelEventHandlerCount(document) is 0
PASS internals.wheelEventHandlerCount(document) is 1
PASS internals.wheelEventHandlerCount(document) is 2
PASS internals.wheelEventHandlerCount(document) is 1
PASS internals.wheelEventHandlerCount(document) is 1
PASS internals.wheelEventHandlerCount(document) is 0
PASS internals.wheelEventHandlerCount(document) is 0
PASS internals.wheelEventHandlerCount(document) is 1
PASS internals.wheelEventHandlerCount(document) is 1
PASS internals.wheelEventHandlerCount(document) is 0
PASS internals.wheelEventHandlerCount(document) is 0
PASS internals.wheelEventHandlerCount(document) is 1
PASS internals.wheelEventHandlerCount(document) is 2
PASS internals.wheelEventHandlerCount(document) is 1
PASS internals.wheelEventHandlerCount(document) is 1
PASS internals.wheelEventHandlerCount(document) is 0
PASS internals.wheelEventHandlerCount(document) is 0
PASS internals.wheelEventHandlerCount(document) is 1
PASS internals.wheelEventHandlerCount(document) is 1
PASS internals.wheelEventHandlerCount(document) is 0
PASS successfullyParsed is true

TEST COMPLETE
